Help Dialog Refactorization
And some additional folder structure improvements
