+++ app/src/main/AndroidManifest.xml
@@ -10,16 +10,16 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.Light.NoActionBar">
 
-        <activity android:name=".StaffActivity"></activity>
-
-        <activity android:name=".MenuActivity"></activity>
+        <activity android:name=".activities.StaffActivity"/>
 
         <activity
-            android:name=".OrderDetailsActivity"
-            android:parentActivityName=".DisplayMenuActivity" />
+            android:name=".activities.MenuActivity"
+            android:parentActivityName=".MainActivity"/>
+
         <activity
-            android:name=".DisplayMenuActivity"
-            android:parentActivityName=".MainActivity" />
+            android:name=".activities.OrderDetailsActivity"
+            android:parentActivityName=".activities.MenuActivity" />
+
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -28,8 +28,8 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".KitchenTableOrder"
-            android:parentActivityName=".StaffActivity"/>
+        <activity android:name=".activities.KitchenTableOrder"
+            android:parentActivityName=".activities.StaffActivity"/>
 </application>
 
 </manifest>
