--- app/src/main/java/com/runtime_terror/myapplication/DisplayMenuActivity.java
@@ -1,27 +1,0 @@
-package com.runtime_terror.myapplication;
-
-import android.content.Intent;
-import android.graphics.Color;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-import android.view.Window;
-import android.view.WindowManager;
-
-public class DisplayMenuActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_display_menu);
-    }
-
-    public void viewOrderDetails(View view){
-        Intent intent = new Intent(this, OrderDetailsActivity.class);
-        startActivity(intent);
-//        Toolbar myToolbar = findViewById(R.id.my_toolbar);
-//        setSupportActionBar(myToolbar);
-    }
-
-}
