+++ app/src/main/java/com/runtime_terror/myapplication/MainActivity.java
@@ -1,10 +1,12 @@
 package com.runtime_terror.myapplication;
 
-import android.app.ActivityManager;
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
+
+import com.runtime_terror.myapplication.activities.MenuActivity;
+import com.runtime_terror.myapplication.activities.StaffActivity;
 
 
 public class MainActivity extends AppCompatActivity {
@@ -13,14 +15,6 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        startActivity(new Intent(getApplicationContext(), DisplayMenuActivity.class));
-        findViewById(R.id.tempStaffSide).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                launchStaffSideActivity(view);
-            }
-        });
     }
 
     public void launchClientSideActivity(View view){
