+++ app/src/main/java/com/runtime_terror/myapplication/adapters/OrderListAdapter.java
@@ -2,22 +2,17 @@ package com.runtime_terror.myapplication.adapters;
 
 import android.content.Context;
 import android.content.Intent;
-import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.animation.AnimationUtils;
 import android.widget.ImageButton;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
-import android.widget.Toast;
 
-import com.runtime_terror.myapplication.KitchenTableOrder;
+import com.runtime_terror.myapplication.activities.KitchenTableOrder;
 import com.runtime_terror.myapplication.R;
-import com.runtime_terror.myapplication.models.Food;
-import com.runtime_terror.myapplication.models.FoodOrder;
 import com.runtime_terror.myapplication.models.Order;
 
 import java.util.Date;
