+++ app/src/main/java/com/runtime_terror/myapplication/interfaces/EditItemInterface.java
@@ -1,4 +1,4 @@
-package com.runtime_terror.myapplication;
+package com.runtime_terror.myapplication.interfaces;
 
 import java.util.List;
