+++ app/src/main/java/com/runtime_terror/myapplication/interfaces/OrderUpdatesListener.java
@@ -1,4 +1,4 @@
-package com.runtime_terror.myapplication;
+package com.runtime_terror.myapplication.interfaces;
 
 public interface OrderUpdatesListener {
     void onOrderUpdate(boolean isComplete);
