--- app/src/main/res/layout/activity_display_menu.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".DisplayMenuActivity">
-
-    <!--<android.support.v7.widget.Toolbar-->
-        <!--android:id="@+id/my_toolbar"-->
-        <!--android:layout_width="match_parent"-->
-        <!--android:layout_height="?attr/actionBarSize"-->
-        <!--android:background="@color/colorPrimary"-->
-        <!--android:elevation="4dp"-->
-        <!--android:theme="@style/ThemeOverlay.AppCompat.ActionBar"-->
-        <!--app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />-->
-
-    <TextView
-        android:id="@+id/tempTextView3"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="This is where the menu will be"
-        android:layout_centerHorizontal="true"
-        android:layout_below="@+id/restName"
-        android:layout_marginTop="60dp"
-        android:textColor="#000000"/>
-
-    <TextView
-        android:id="@+id/restName"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Restaurant Name"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="60dp"
-        android:layout_below="?attr/actionBarSize"
-        android:textColor="#000000"/>
-
-    <ImageButton
-        android:id="@+id/placeholderButton"
-        android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:onClick="viewOrderDetails"
-        android:src="@drawable/ic_cart_icon_red"
-        android:layout_alignParentEnd="true"
-        android:layout_marginTop="60dp"
-        android:layout_marginRight="10dp"
-        android:backgroundTint="#ffffff"/>
-</RelativeLayout>
