+++ app/src/main/res/layout/activity_kitchen_table_order.xml
@@ -8,7 +8,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".KitchenTableOrder"
+    tools:context=".activities.KitchenTableOrder"
     android:background="@drawable/food_card_bg"
     >
     <android.support.v7.widget.Toolbar
