+++ app/src/main/res/layout/activity_order_details.xml
@@ -10,7 +10,7 @@
         android:layout_height="match_parent"
         android:background="@color/white"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        tools:context=".OrderDetailsActivity">
+        tools:context=".activities.OrderDetailsActivity">
 
         <android.support.v7.widget.Toolbar
             android:id="@+id/toolbar"
