+++ app/src/main/res/layout/activity_staff.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StaffActivity">
+    tools:context=".activities.StaffActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbar"
